Use wtfrost instead of blockstack-core and fix tests
